ICU-13171 Fix issues with ICU4C Samples, and various issues with vcxproj files. The samples now all build on Win32, except for the layout sample (but the LE is deprecated).
X-SVN-Rev: 40951

+++ icu4c/source/samples/ugrep/ugrep.cpp
@@ -291,7 +291,7 @@ void readFile(const char *name) {
     //   Read in the file
     //
     charBuf    = (char *)realloc(charBuf, rawFileLen+1);   // Need error checking...
-    int t = fread(charBuf, 1, rawFileLen, file);
+    int t = static_cast<int>(fread(charBuf, 1, rawFileLen, file));
     if (t != rawFileLen)  {
         fprintf(stderr, "Error reading file \"%s\"\n", fileName);
         fclose(file);
